Fix tests by still handling the case that the user passes in a raw ioredis instance.
