+++ src/__tests__/ioredis-passed-in.test.ts
